remove Travis black check
At the moment this check is very much version dependant: Travis installs by default an earlier version "black-19.3b0" to the one on my computer "black-19.10b0", and this causes the build to fail. The two versions have different formatting definitions. I'll wait until it exits the beta stage to do this check