Add arbitrary non linear inductive dipole and third order terms. Add function returning the full hamiltonian with sympy.
Allows to insert an arbitrary non linear inductive dipole (class D), taking into account at least orders 2 to 4 terms in the inductive energy expansion in series.
It also calculates three waves mixing coefficients between modes to also get third order terms in the hamiltonian, in the same way than it is currently performed for anharmonicities.
The dipole is represented as a 'D' in a box in the GUI editor and as a black triangle in matplotlib.
A new function 'hamiltonian_sym' returns the full hamiltonian using sympy.physics.secondquant library.
The security preventing the user from giving too small or negative labels has been removed since higher order terms in the energy may be negative.

The arbitrary dipole can't have determined values, it has to be fully labelized.